Test/test_bagtile.py
import unittest

from Game.bagtile import BagTiles
from Game.tile import Tile

from unittest.mock import patch

class TestBagTiles(unittest.TestCase):
    @patch('random.shuffle')
    def test_bag_tiles(self, patch_shuffle):
        bag = BagTiles()
        self.assertEqual(len(bag.tiles), 100)
        self.assertEqual(patch_shuffle.call_count, 1)
        self.assertEqual(patch_shuffle.call_args[0][0], bag.tiles)

    def test_take(self):
        bag = BagTiles()
        tiles = bag.take(2)
        self.assertEqual(len(tiles), 2)
        self.assertEqual(len(bag.tiles), 100 - 2)

    def test_take_invalid(self):
        bag = BagTiles()
        with self.assertRaises(ValueError):
            tiles = bag.take(101)

    def test_put(self):
        bag = BagTiles()
        put_tiles = [Tile('Z', 1), Tile('Y', 1)]
        bag.put(put_tiles)
        self.assertEqual(len(bag.tiles),102,)

if __name__ == '__main__':
    unittest.main()